Use GraphQL client from NuGet instead of reinventing the wheel
This commit introduces Client.cs and EngineConnection.cs which are translations of the corresponding files from the TypeScript SDK.  They are placed in a separate csproj from the SDK proper so that they can be a dependency of the generator and the SDK. If the generator requires the SDK, then it will be troublesome to re-run the generator whenever it has produced compilation errors.

This commit knowingly breaks the DaggerSDK csproj, you will find that it is fixed in a following commit on this branch.