Spawn Dagger executable from SDK Generator
A portion of bin.ts is translated for this purpose; the download part will be separated to static functions in the primary SDK's assembly. The abstractions can be improved (i.e. existing-session-based connection should also be an implementation of IEngineConnection), but until it's 100% translated and tested, I'm keeping things as 1:1 with the TypeScript as practical.