Encapsulate serialization logic
Before, as with the TypeScript SDK from which this is based, ComputeNestedQueries mutates parameters which are then passed to BuildQuery - the latter expects the former to have been run on the given data.  This sort of "trust fall" is brittle and inexpressive.  Additionally, in an OO language like this, encapsulating type-specific logic in classes that reify those types is preferable to loose coordination through enums and other conventions.  So, parameter serialization logic has been encapsulated into composable per-serialization-strategy implementations of a parameter value class.  The pattern of dispatching queries in parallel before serializing them in series is retained, but with a failsafe that results in the "priming" step being purely an optimization rather than a necessity.  In addition to the general maintainability merits, the primary motivation behind this change is making it significantly easier to cleanly implement lazy evaluation of enumerable arguments.