Directly compile generated syntax tree to Assembly instead of source code.
For Python and TypeScript, there is no option other than to create the source code.  For other languages, generating a binary is bad because it is useful to be able to debug it. With dotnet, thanks to the brilliance of CIL and robustness of its SDK tooling, a `.dll` is hardly any less convenient than a `.cs`.